fall_detection_ws/requirements.txt
fastapi
uvicorn[standard]
websockets
tensorflow
opencv-python
numpy
Pillow
